Permit spaces in <select> content.
Without this, markup such as <select>Foo bar</select> causes an
attempt to match a[data-dk-dropdown-value=Foo Bar], which fails.

Stop-gap fix is to wrap quotes around the value.  This obviously
isn't ideal, since there is no guarantee that the value itself
doesn't already contain quotes which would then need escaping.
However, testing has shown DropKick is already broken in that case,
so this patch at least doesn't make it _more_ broken.

However, more work needs to be done to come up with a clean solution
that allows for various 'special' characters in the contained
markup.